Better organize metrics into sections
