+++ Demo/Resources/Localizable.xcstrings
@@ -61,6 +61,9 @@
     "Application" : {
 
     },
+    "Asset loading" : {
+
+    },
     "Audiovisual background policy" : {
 
     },
@@ -151,6 +154,12 @@
     "Indicated bitrate (Mbps)" : {
 
     },
+    "Information" : {
+
+    },
+    "Initial buffering" : {
+
+    },
     "Kind" : {
 
     },
@@ -223,6 +232,9 @@
     "Play" : {
 
     },
+    "Playback duration" : {
+
+    },
     "Playback HUD" : {
 
     },
@@ -242,6 +254,9 @@
 
     },
     "Reset" : {
+
+    },
+    "Resource loading" : {
 
     },
     "Search" : {
@@ -280,6 +295,9 @@
     "Stalls" : {
 
     },
+    "Startup times" : {
+
+    },
     "Stop" : {
 
     },
@@ -305,6 +323,12 @@
 
     },
     "Tracking" : {
+
+    },
+    "Type" : {
+
+    },
+    "URI" : {
 
     },
     "URL" : {
